scripts/training/selim/multiclass-lightning/requirements.txt
transformers==4.8.2
tensorflow==2.4.0
pytorch-lightning==1.3.8
torchmetrics==0.4.1
tqdm==4.41.1
nlpaug==1.1.6
nltk==3.2.5

